Fix issue with Gradle 8.6 and catalog access (projects without catalog)
In earlier version, in a standard configuration, if the
'VersionCatalogsExtension' existed, there was also a catalog called
'libs'. If there were no catalog at all, the extension did not
exist. This has changed recently. This causes the plugin to fail
in a standard setup where the catalog is NOT used.